Initial rough draft of value struct support
