+++ cs-bindgen-macro/src/enumeration.rs
@@ -1,4 +1,4 @@
-use crate::func::*;
+use crate::{reject_generics, value};
 use proc_macro2::{Literal, TokenStream};
 use quote::*;
 use syn::*;
@@ -152,35 +152,7 @@ fn quote_complex_enum(item: &ItemEnum) -> syn::Result<TokenStream> {
             return None;
         }
 
-        // Extract the list of fields for the binding struct. The generated struct is the
-        // same for both struct-like and tuple-like variants, though in the latter case we
-        // have to manually generate names for the fields based on the index of the element.
-        let from_fields = variant
-            .fields
-            .iter()
-            .enumerate()
-            .map(|(index, field)| {
-                let field_ty = &field.ty;
-                let field_ident = field
-                    .ident
-                    .as_ref()
-                    .map(Clone::clone)
-                    .unwrap_or_else(|| format_ident!("element_{}", index));
-
-                quote! {
-                    #field_ident: <#field_ty as cs_bindgen::abi::Abi>::Abi
-                }
-            })
-            .collect::<Vec<_>>();
-
-        Some(quote! {
-            #[repr(C)]
-            #[derive(Clone, Copy)]
-            #[allow(bad_style)]
-            pub struct #abi_ident {
-                #( #from_fields, )*
-            }
-        })
+        Some(value::quote_abi_struct(&abi_ident, &variant.fields))
     });
 
     let abi_union_fields = item.variants.iter().filter_map(|variant| {
